Refactor thresholdColorField function and move it to bulletDataUtils for better organization
